S01E02/requirements.txt
openai
requests
python-dotenv
loguru